+++ agent/models/src/shard.rs
@@ -1,19 +1,61 @@
+/// Information about the current commit offset of a shard or replication lag.
+#[derive(Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Debug, Serialize, Deserialize)]
+pub struct CommitOffset {
+    pub unit: CommitUnit,
+    pub value: i64,
+}
+
+impl CommitOffset {
+    pub fn new(value: i64, unit: CommitUnit) -> CommitOffset {
+        CommitOffset { unit, value }
+    }
+
+    pub fn seconds(value: i64) -> CommitOffset {
+        CommitOffset::new(value, CommitUnit::seconds())
+    }
+
+    pub fn unit<S: Into<String>>(value: i64, unit: S) -> CommitOffset {
+        CommitOffset::new(value, CommitUnit::unit(unit))
+    }
+}
+
+
+/// Unit of commit offsets or replica lags.
+#[derive(Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Debug, Serialize, Deserialize)]
+pub enum CommitUnit {
+    Seconds,
+    Unit(String),
+}
+
+impl CommitUnit {
+    pub fn seconds() -> CommitUnit {
+        CommitUnit::Seconds
+    }
+
+    pub fn unit<S: Into<String>>(unit: S) -> CommitUnit {
+        CommitUnit::Unit(unit.into())
+    }
+}
+
+
 /// Information about a shard on a node.
 #[derive(Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Debug, Serialize, Deserialize)]
 pub struct Shard {
+    pub commit_offset: Option<CommitOffset>,
     pub id: String,
+    pub lag: Option<CommitOffset>,
     pub role: ShardRole,
-    pub lag: Option<i64>,
-    pub last_op: i64,
 }
 
 impl Shard {
-    pub fn new<S>(id: S, role: ShardRole, lag: Option<i64>, last_op: i64) -> Shard
+    pub fn new<S>(
+        id: S, role: ShardRole, commit_offset: Option<CommitOffset>, lag: Option<CommitOffset>
+    ) -> Shard
         where S: Into<String>,
     {
         Shard {
             id: id.into(),
-            role, lag, last_op,
+            commit_offset, lag, role,
         }
     }
 }
@@ -44,31 +86,51 @@ pub enum ShardRole {
 #[cfg(test)]
 mod tests {
     use serde_json;
+    use super::CommitOffset;
     use super::Shard;
     use super::ShardRole;
 
     #[test]
     fn primary_from_json() {
-        let payload = r#"{"id":"shard-1","role":"Primary","lag":0,"last_op":12345}"#;
+        let payload = concat!(
+            r#"{"commit_offset":{"unit":"Seconds","value":12345},"id":"shard-1","#,
+            r#""lag":{"unit":"Seconds","value":0},"role":"Primary"}"#
+        );
         let shard: Shard = serde_json::from_str(payload).unwrap();
-        let expected = Shard::new("shard-1", ShardRole::Primary, Some(0), 12345);
+        let expected = Shard::new(
+            "shard-1", ShardRole::Primary,
+            Some(CommitOffset::seconds(12345)),
+            Some(CommitOffset::seconds(0))
+        );
         assert_eq!(shard, expected);
     }
 
     #[test]
     fn primary_to_json() {
-        let shard = Shard::new("shard-1", ShardRole::Primary, Some(0), 12345);
+        let shard = Shard::new(
+            "shard-1", ShardRole::Primary,
+            Some(CommitOffset::seconds(12345)),
+            Some(CommitOffset::seconds(0))
+        );
         let payload = serde_json::to_string(&shard).unwrap();
-        let expected = r#"{"id":"shard-1","role":"Primary","lag":0,"last_op":12345}"#;
+        let expected = concat!(
+            r#"{"commit_offset":{"unit":"Seconds","value":12345},"id":"shard-1","#,
+            r#""lag":{"unit":"Seconds","value":0},"role":"Primary"}"#
+        );
         assert_eq!(payload, expected);
     }
 
     #[test]
     fn unkown_from_json() {
-        let payload = r#"{"id":"shard-1","role":{"Unknown":"Test"},"lag":0,"last_op":12345}"#;
+        let payload = concat!(
+            r#"{"commit_offset":{"unit":"Seconds","value":12345},"id":"shard-1","#,
+            r#""lag":{"unit":"Seconds","value":0},"role":{"Unknown":"Test"}}"#
+        );
         let shard: Shard = serde_json::from_str(payload).unwrap();
         let expected = Shard::new(
-            "shard-1", ShardRole::Unknown(String::from("Test")), Some(0), 12345
+            "shard-1", ShardRole::Unknown(String::from("Test")),
+            Some(CommitOffset::seconds(12345)),
+            Some(CommitOffset::seconds(0))
         );
         assert_eq!(shard, expected);
     }
@@ -76,34 +138,99 @@ mod tests {
     #[test]
     fn unkown_to_json() {
         let shard = Shard::new(
-            "shard-1", ShardRole::Unknown(String::from("Test")), Some(0), 12345
+            "shard-1", ShardRole::Unknown(String::from("Test")),
+            Some(CommitOffset::seconds(12345)),
+            Some(CommitOffset::seconds(0))
         );
         let payload = serde_json::to_string(&shard).unwrap();
-        let expected = r#"{"id":"shard-1","role":{"Unknown":"Test"},"lag":0,"last_op":12345}"#;
+        let expected = concat!(
+            r#"{"commit_offset":{"unit":"Seconds","value":12345},"id":"shard-1","#,
+            r#""lag":{"unit":"Seconds","value":0},"role":{"Unknown":"Test"}}"#
+        );
         assert_eq!(payload, expected);
     }
 
     #[test]
-    fn missing_lag_from_json() {
-        let payload = r#"{"id":"shard-1","role":"Secondary","last_op":12345}"#;
+    fn missing_commit_offset_from_json() {
+        let payload = concat!(
+            r#"{"id":"shard-1","#,
+            r#""lag":{"unit":"Seconds","value":0},"role":{"Unknown":"Test"}}"#
+        );
         let shard: Shard = serde_json::from_str(payload).unwrap();
-        let expected = Shard::new("shard-1", ShardRole::Secondary, None, 12345);
+        let expected = Shard::new(
+            "shard-1", ShardRole::Unknown(String::from("Test")),
+            None, Some(CommitOffset::seconds(0))
+        );
+        assert_eq!(shard, expected);
+    }
+
+    #[test]
+    fn missing_lag_from_json() {
+        let payload = concat!(
+            r#"{"commit_offset":{"unit":"Seconds","value":12345},"id":"shard-1","#,
+            r#""role":{"Unknown":"Test"}}"#
+        );
+        let shard: Shard = serde_json::from_str(payload).unwrap();
+        let expected = Shard::new(
+            "shard-1", ShardRole::Unknown(String::from("Test")),
+            Some(CommitOffset::seconds(12345)), None
+        );
+        assert_eq!(shard, expected);
+    }
+
+    #[test]
+    fn no_commit_offset_from_json() {
+        let payload = concat!(
+            r#"{"commit_offset":null,"id":"shard-1","#,
+            r#""lag":{"unit":"Seconds","value":0},"role":{"Unknown":"Test"}}"#
+        );
+        let shard: Shard = serde_json::from_str(payload).unwrap();
+        let expected = Shard::new(
+            "shard-1", ShardRole::Unknown(String::from("Test")),
+            None, Some(CommitOffset::seconds(0))
+        );
         assert_eq!(shard, expected);
     }
 
     #[test]
     fn no_lag_from_json() {
-        let payload = r#"{"id":"shard-1","role":"Secondary","lag":null,"last_op":12345}"#;
+        let payload = concat!(
+            r#"{"commit_offset":{"unit":"Seconds","value":12345},"id":"shard-1","#,
+            r#""lag":null,"role":{"Unknown":"Test"}}"#
+        );
         let shard: Shard = serde_json::from_str(payload).unwrap();
-        let expected = Shard::new("shard-1", ShardRole::Secondary, None, 12345);
+        let expected = Shard::new(
+            "shard-1", ShardRole::Unknown(String::from("Test")),
+            Some(CommitOffset::seconds(12345)), None
+        );
         assert_eq!(shard, expected);
     }
 
     #[test]
-    fn no_lag_to_json() {
-        let shard = Shard::new("shard-1", ShardRole::Primary, None, 12345);
+    fn no_commit_offset_to_json() {
+        let shard = Shard::new(
+            "shard-1", ShardRole::Primary,
+            None, Some(CommitOffset::seconds(0))
+        );
         let payload = serde_json::to_string(&shard).unwrap();
-        let expected = r#"{"id":"shard-1","role":"Primary","lag":null,"last_op":12345}"#;
+        let expected = concat!(
+            r#"{"commit_offset":null,"id":"shard-1","#,
+            r#""lag":{"unit":"Seconds","value":0},"role":"Primary"}"#
+        );
+        assert_eq!(payload, expected);
+    }
+
+    #[test]
+    fn no_lag_to_json() {
+        let shard = Shard::new(
+            "shard-1", ShardRole::Primary,
+            Some(CommitOffset::seconds(12345)), None
+        );
+        let payload = serde_json::to_string(&shard).unwrap();
+        let expected = concat!(
+            r#"{"commit_offset":{"unit":"Seconds","value":12345},"id":"shard-1","#,
+            r#""lag":null,"role":"Primary"}"#
+        );
         assert_eq!(payload, expected);
     }
 }